backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
attrs==25.3.0
av==15.0.0
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.5
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
docstring_parser==0.16
eval_type_backport==0.2.2
flatbuffers==25.2.10
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jiter==0.10.0
jwt==1.4.0
livekit==1.0.9
livekit-agents==1.1.4
livekit-api==1.0.2
livekit-plugins-deepgram==1.1.4
livekit-plugins-elevenlabs==1.1.4
livekit-plugins-google==1.1.4
livekit-plugins-openai==1.1.4
livekit-plugins-silero==1.1.4
livekit-protocol==1.0.3
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.2.1
onnxruntime==1.22.0
openai==1.94.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyreadline3==3.5.4
python-dotenv==1.1.1
sniffio==1.3.1
sounddevice==0.5.2
sqlparse==0.5.3
sympy==1.14.0
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1